Struct processor::Processor

pub struct Processor { /* private fields */ }

Implementations

impl Processor

pub fn new() -> Self

Cria um novo processador com NUM_REGISTERS registradores, memória de tamanho MEMORY_SIZE e mais 4 registradores especiais que são: Flag Register (FR), Program Counter (PC), Stack Pointer (SP) e Instruction Register (IR). Inicialmente, todos os endereços da memória e todos os registradores guardam o valor 0x0.

pub fn new_debug(s: usize) -> Self

Cria um novo processador para DEBUG com NUM_REGISTERS registradores, memória de tamanho s e mais 4 registradores especiais que são: Flag Register (FR), Program Counter (PC), Stack Pointer (SP) e Instruction Register (IR). Inicialmente, todos os endereços da memória e todos os registradores guardam o valor 0x0.

Importante

Deve ser utilizado somente para testes!

pub fn arc_mem(&self) -> Arc<Mutex<Vec<MemoryCell>>>

pub fn mem(&self, addr: MemoryCell) -> Result<MemoryCell, ProcessorError>

Retorna o valor presente no endereço addr da memória.

Erros

ProcessorError::InvalidAddress caso o endereço seja inválido.

ProcessorError::Generic em caso de std::sync::PoisonError.

Exemplo

use crate::processor::Processor;

let mut p = Processor::new();
assert_eq!(0x0, p.mem(0).unwrap());

pub fn set_mem( &mut self, adrr: MemoryCell, v: MemoryCell ) -> Result<(), ProcessorError>

Altera o valor salvo no endereço addr para v.

Erros

ProcessorError::InvalidAddress caso o endereço seja inválido.

ProcessorError::Generic em caso de std::sync::PoisonError.

Exemplo

use crate::processor::Processor;

let mut p = Processor::new();
assert_eq!(0x0, p.mem(0).unwrap());
p.set_mem(0, 0x1);
assert_eq!(0x1, p.mem(0).unwrap());

pub fn reg(&self, n: MemoryCell) -> Result<MemoryCell, ProcessorError>

Retorna o valor do registrador n.

Erros

Retorna o erro ProcessorError::InvalidRegister caso o registrador n seja inválido.

Exemplo

use crate::processor::Processor;

let mut p = Processor::new();
assert_eq!(0x0, p.reg(0).unwrap());

pub fn set_reg( &mut self, n: MemoryCell, v: MemoryCell ) -> Result<(), ProcessorError>

Esta função altera o conteúdo do registrar n para v.

Erros

Retorna o erro ProcessorError::InvalidRegister caso o registrador n seja invalido.

Exemplo

use crate::processor::Processor;

let mut p = Processor::new();
assert_eq!(0x0, p.reg(0).unwrap());
p.set_reg(0, 0x1);
assert_eq!(0x1, p.reg(0).unwrap());

pub fn rx(&self) -> MemoryCell

Retorna o valor presente no campo RX do registrador IR.

Mapeamento

Nº	15	14	13	12	11	10	09	08	07	06	05	04	03	02	01	00

IR	–	–	–	–	–	–	RX	RX	RX	–	–	–	–	–	–	–

pub fn ry(&self) -> MemoryCell

Retorna o valor presente no campo RY do registrador IR.

Mapeamento

Nº	15	14	13	12	11	10	09	08	07	06	05	04	03	02	01	00

IR	–	–	–	–	–	–	–	–	–	RY	RY	RY	–	–	–	–

pub fn rz(&self) -> MemoryCell

Retorna o valor presente no campo RZ do registrador IR.

Mapeamento

Nº	15	14	13	12	11	10	09	08	07	06	05	04	03	02	01	00

IR	–	–	–	–	–	–	–	–	–	–	–	–	RZ	RZ	RZ	–

pub fn fr(&self, i: usize) -> Result<bool, ProcessorError>

Retorna o valor da i-ésima flag do registrador Flag Register.

⚠️ Atenção

É recomendado utilizar os índices mapeados em [isa::FlagIndex] para evitar erros e comportamentos indesejados.

Erros

Retorna o erro ProcessorError::InvalidFlag caso o valor i seja inválido.

pub fn set_fr(&mut self, i: usize, v: bool) -> Result<(), ProcessorError>

Altera o valor do bit i do registrador Flag Register para o valor v.

⚠️ Atenção

É recomendado utilizar os índices mapeados em [isa::FlagIndex] para evitar erros e comportamentos indesejados.

Erros

Retorna o erro ProcessorError::InvalidFlag caso o valor i seja inválido.

pub fn ir(&self) -> MemoryCell

Retorna o valor do registrador especial Instruction Register.

pub fn sp(&self) -> MemoryCell

Retorna o valor do registrador especial Stack Pointer.

pub fn set_sp(&mut self, addr: MemoryCell) -> Result<(), ProcessorError>

Altera o valor do registrador especial Stack Pointer para addr.

Erros

ProcessorError::StackOverflow caso o addr seja menor que o topo da pilha.

ProcessorError::StackUnderflow caso o addr seja maior que a base da pilha.

pub fn inc_sp(&mut self, v: MemoryCell) -> Result<(), ProcessorError>

Incrementa o valor do registrador especial Stack Pointer de um valor v.

Erros

ProcessorError::StackOverflow caso o resultado seja menor que o topo da pilha.

ProcessorError::StackUnderflow caso o resultado seja maior que a base da pilha.

pub fn dec_sp(&mut self, v: usize) -> Result<(), ProcessorError>

Decrementa o valor do registrador especial Stack Pointer de um valor v.

Erros

ProcessorError::StackOverflow caso o resultado seja menor que o topo da pilha.

ProcessorError::StackUnderflow caso o resultado seja maior que a base da pilha.

ProcessorError::InvalidAddress caso o resultado seja negativo.

pub fn pc(&self) -> MemoryCell

Retorna o valor do registrador especial Program Counter.

pub fn set_pc(&mut self, addr: usize) -> Result<(), ProcessorError>

pub fn inc_pc(&mut self, v: usize) -> Result<(), ProcessorError>

Incrementa o valor do registrador especial Program Counter de um valor v.

Erros

Retorna ProcessorError::SegmentationFault caso o resultado esteja fora da área de programa e variáveis.

pub fn set_halted(&mut self, value: bool)

pub fn halted(&self) -> bool

pub fn set_debug(&mut self, value: bool)

pub fn debug(&self) -> bool

pub fn set_pixel(&mut self, value: Option<(Pixelmap, usize)>)

pub fn pixel(&mut self) -> Option<(Pixelmap, usize)>

pub fn ula_operation(&mut self) -> Result<(), ProcessorError>

pub fn next(&mut self) -> Result<(), ProcessorError>